requirements.txt
pycryptodome
Flask
requests
crypto-otrs